requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
openai==0.28.1
chromadb==0.4.18
numpy<2.0.0  # Required for ChromaDB compatibility
